Move sieve scripts to home directory
This moves the sieve scripts from ea separate directory into the home
directory of each user. The sieve scripts will be in a directory next to
the maildir.
This is required for setups that do not use email addresses as their login
name, as there is no %d available and %u might be different depending on,
whether new email is processed or a user session is processed.

Attention: This change requires the administrator to move all existing
sieve scripts to the new location.