Update Postfix to 3.5 for X-Original-To header with LMTP
Easiest way: Update base of Postfix image to Debian Bullseye, as it
contains a sufficiently recent version of Postfix.

Postfix added this with changelog entry 20190615.

The flags DORX are added to the lmtp process to enable the new features.

Also Dovecot is told about the new X-Original-To header.

This is the configuration file named after LDA, but we use LMTP. My
understanding is, that this option is used by the LMTP port of Dovecot as
well.